Move hook execution tracer to tracing module
